requirements.txt
streamlit==1.36.0
pytesseract==0.3.10
Pillow==10.3.0
openai==1.35.10
python-dotenv==1.0.1
